+++ app/database/schema/app/table_schemas/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export {default as app} from './app';
+export {default as info} from './info';
 export {default as global} from './global';
 export {default as servers} from './servers';
