+++ app/database/schema/app/test.ts
@@ -39,16 +39,16 @@ describe('*** Test schema for APP database ***', () => {
                     columns: {
                         db_path: {name: 'db_path', type: 'string'},
                         display_name: {name: 'display_name', type: 'string'},
-                        url: {name: 'url', type: 'string', isIndexed: true},
-                        last_active_at: {name: 'last_active_at', type: 'number', isIndexed: true},
                         identifier: {name: 'identifier', type: 'string', isIndexed: true},
+                        last_active_at: {name: 'last_active_at', type: 'number', isIndexed: true},
+                        url: {name: 'url', type: 'string', isIndexed: true},
                     },
                     columnArray: [
                         {name: 'db_path', type: 'string'},
                         {name: 'display_name', type: 'string'},
-                        {name: 'url', type: 'string', isIndexed: true},
-                        {name: 'last_active_at', type: 'number', isIndexed: true},
                         {name: 'identifier', type: 'string', isIndexed: true},
+                        {name: 'last_active_at', type: 'number', isIndexed: true},
+                        {name: 'url', type: 'string', isIndexed: true},
                     ],
                 },
             },
